requirements.txt
fastapi[standard]
pdfplumber
google-genai
python-multipart
fuzzywuzzy
python-dotenv
scikit-learn
streamlit
